+++ strato/CHANGELOG.md
@@ -19,6 +19,7 @@ so that they could be properly moved to their respective version's subsection.
 - `strictDecimal` pragma for decimal precision strictness
 - `truncate(uint)` built-in method for decimal numbers
 - Added `pragma safeExternalCalls` for contracts that want to enforce extra type safety on external calls from other contracts
+- Added `pragma solidvm 11.4` that includes all existing pragmas and their features
 
 ### Changed